use bytecode search in all parts. fix tests
