+++ flow-maven-plugin/src/main/java/com/vaadin/flow/plugin/maven/NodeUpdateImportsMojo.java
@@ -31,7 +31,6 @@ import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.plugins.annotations.ResolutionScope;
 
-import com.vaadin.flow.server.frontend.AnnotationValuesExtractor;
 import com.vaadin.flow.server.frontend.FrontendToolsLocator;
 import com.vaadin.flow.server.frontend.NodeUpdateImports;
 import com.vaadin.flow.server.frontend.NodeUpdater;
@@ -57,9 +56,7 @@ public class NodeUpdateImportsMojo extends NodeUpdateAbstractMojo {
     @Override
     protected NodeUpdater getUpdater() {
         if (updater == null) {
-            AnnotationValuesExtractor extractor = new AnnotationValuesExtractor(
-                    getClassFinder(project));
-            updater = new NodeUpdateImports(extractor, jsFile, npmFolder,
+            updater = new NodeUpdateImports(getClassFinder(project), jsFile, npmFolder,
                     nodeModulesPath, convertHtml);
         }
         return updater;
