+++ flow-maven-plugin/src/main/java/com/vaadin/flow/plugin/maven/NodeUpdatePackagesMojo.java
@@ -23,7 +23,6 @@ import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.plugins.annotations.ResolutionScope;
 
-import com.vaadin.flow.server.frontend.AnnotationValuesExtractor;
 import com.vaadin.flow.server.frontend.NodeUpdatePackages;
 import com.vaadin.flow.server.frontend.NodeUpdater;
 
@@ -46,13 +45,11 @@ public class NodeUpdatePackagesMojo extends NodeUpdateAbstractMojo {
     @Override
     protected NodeUpdater getUpdater() {
         if (updater == null) {
-            AnnotationValuesExtractor extractor = new AnnotationValuesExtractor(
-                    getClassFinder(project));
             File webpackOutputRelativeToProjectDir = project.getBasedir()
                     .toPath().relativize(getWebpackOutputDirectory().toPath())
                     .toFile();
 
-            updater = new NodeUpdatePackages(extractor,
+            updater = new NodeUpdatePackages(getClassFinder(project),
                     webpackOutputRelativeToProjectDir, webpackTemplate,
                     npmFolder, nodeModulesPath, convertHtml);
         }
